+++ src/core/debug/measure/get-input-fn.js
@@ -3,17 +3,26 @@
  * @author BBC Children's D+E
  * @license Apache-2.0
  */
+const throttleFor = 100;
+let lastCallTime = 0;
+
 export const getInputFn = keys => () => {
-    const ctrl = keys.ctrl.isDown;
-    const val = keys.shift.isDown ? 10 : 1;
+    const currentTime = Date.now();
+    const sizing = keys.x.isDown;
+    const slow = keys.z.isDown;
+    const throttled = slow && currentTime - lastCallTime < throttleFor;
+
+    const val = throttled ? 0 : slow ? 1 : 10;
 
     const xDir = keys.right.isDown - keys.left.isDown;
     const yDir = keys.down.isDown - keys.up.isDown;
 
-    const x = ctrl ? 0 : xDir * val;
-    const y = ctrl ? 0 : yDir * val;
-    const width = ctrl ? xDir * val : 0;
-    const height = ctrl ? yDir * val : 0;
+    const x = sizing ? 0 : xDir * val;
+    const y = sizing ? 0 : yDir * val;
+    const width = sizing ? xDir * val : 0;
+    const height = sizing ? yDir * val : 0;
+
+    !throttled && (lastCallTime = currentTime);
 
     return { x, y, width, height };
 };